Create automation_menus.py
Initial commit for automation_menus.py which currently has the set_prefix script attached as a menu option. The script needs to be checked to make sure it's not botching names but also it should possibly move items besides post process materials into their own directories, if that wouldn't be too disruptive.